New ultimate AI fo red.
Can only loose in small map, 7 enemies, more resources setting,
but only if all its villages are captured in the beginning.
So usually wins, but not from starter village in this config.
The nature of this config is that winning heavily depends on luck.

In every other config it won every single time, when facing old AI (at any difficulty).